+++ packages/shared/common/src/internal/fdv2/proto.ts
@@ -6,18 +6,32 @@ export type EventType =
   | 'goodbye'
   | 'error'
   | 'heart-beat';
+
 export type IntentCode = 'xfer-full' | 'xfer-changes' | 'none';
 export type ObjectKind = 'flag' | 'segment';
 
-export interface Event {
-  event: EventType;
+export interface FDv2Event {
+  /**
+   * The event type could be one we know, or it could be any string.
+   * This is for forward compatibility and to make it clear the protocol may send us types we don't recognize.
+   */
+  event: EventType | string;
+  /**
+   * Could be one of many known types, or an unknown type.
+   * The unknown type is to handle forward compatibility.
+   */
   data:
     | ServerIntentData
     | PutObject
     | DeleteObject
     | PayloadTransferred
     | GoodbyeObject
-    | ErrorObject;
+    | ErrorObject
+    | unknown;
+}
+
+export interface FDv2EventsCollection {
+  events: FDv2Event[];
 }
 
 export interface ServerIntentData {